Act070_Main  - Params para serem usado no bundle  - Add var de tradução  - Modificado getAct071Bundle add param que identicica se tem ou não criação de action  - Modificado metodo setCheckinToStep, add update required no step  - Modificado novamente defineActionFlow,MAS AINDA NÃO ESTA OK. Tratado status done do step e ctrl, mas falta tratar outros casos  - Modificado metodos addCheckinCtrl, addNewProcess, addCheckOutCtrl, setando o stepCode no obj @#$ Act071  - Add var que indicam a criação de ctrl e action  - Add set das novas var no recebimento do bundle  - Modificado metodo updateActionData, para chamar metodo que seta dados de inicio no controle, caso não existam, e metodo que cria a action do ctrl caso seja processo d criação de action  - Add validação de se action é null antes de chamar os metodo setReadOnly e  setDataToViews.  - Modificado metodo setDataToObj, p setar o nome da foto no campo setAction_photo_name, pois agora esse é o campo que o server receberá a foto.  - Criado metodo setStartInfoIfNeed que seta os dados de start_* caso o status do ctl seja pending e usr code null  - Criado metodo createActionIfNeed, que cria action para o control, caso a flag isCreationAction seja verdadeira.  - Modificado todo metodo updateTicketAction, para setar o ctrl no step e atualizar os dados do step e update required no ticket  - Modificado o metodo getCtrlIdx para varer os ctrls do step e achar o indice da substituição - VERIFICAR COMO FAZER QUANDO FOR UM NOVO  - Criado metodo getStepIdx que resgata do ticket o step do ctrl @#$ DatabaseHelperMulti,TK_Ticket_Product, TK_Ticket_ProductDao  -  Removido update_required da tabela tk_ticket_product @#$ Sql_WS_TK_Ticket_Save_001  - Removido left join com produto, pois agora será verificado o  update_required_product no cabeçalho @#$ Sql_WS_TK_Ticket_Save_002  - Query que seta update required e update_required_product no cabeçalho do ticket @#$ Sql_WS_TK_Ticket_Save_003  -  Query que seta update required no step @#$ Sql_WS_TK_Ticket_Save_004  -  Query que seta update required nos ctrls @#$ StepProcessBtn  - Add atributo stepCode no construtor @#$ T_TK_Ticket_Save_Env  - Add notação de expose no atributo ticket @#$ T_TK_Ticket_Save_Rec - Ainda em dev  - Modifiado modelo de T_TK_Ticket_Save_Rec add lista de ticket e from_to @#$ T_TK_Ticket_Save_Rec_From_To  - Criado modelo baseado no retorno recebeido ATÉ AGORA - Em dev @#$ T_TK_Ticket_Save_Rec_Result  - Modificado modelo  - iniciando processamento do de-para @#$ T_TK_Ticket_Save_Rec_Result_Step  - Criado modelo para atributo step do obj retornado @#$ TK_Ticket_Action  - Add notação de expose no atributo action_photo_name @#$ TK_Ticket_Ctrl  - Criado metodo setPKIntoProcess que seta a pks nos itens filhos @#$ TK_Ticket_Ctrl_Sql_004  -  Query que seleciona o proximo ticket_seq_tmp para ser usado no ctrl e processos filhos @#$ TK_Ticket_CtrlDao  - Modificado metodo deleteByCtrlType passando como param o ctrl e assim valiando obj é null antes de tentar o delete - REVER POIS SE MUDAR DE VARIS ITENS PARA NENHUM, NÃO FUNCIONARÁ DEVIDAMENTE @#$ TK_Ticket_Step  - Modificado  setPK para usar o this o inves do tk_ticket no set das infos do ctrl @#$ WS_TK_Ticket_Download  - Modificado metodo de download para fazer um removeFull do obj antes de atual de atualizar @#$ WS_TK_Ticket_Save - ainda em dev  - Iniciado processamento do save  - Envio aparentemente ok  - De para ´so jesus na causa @#$

+++ app/src/main/java/com/namoadigital/prj001/dao/TK_Ticket_CtrlDao.java
@@ -395,7 +395,7 @@ public class TK_Ticket_CtrlDao extends BaseDao implements DaoWithReturn<TK_Ticke
                 sbWhere.append(" and ");
                 sbWhere.append(STEP_CODE).append(" = '").append(tk_ticket_ctrl.getStep_code()).append("'");
                 //Tenta o delete do tipo do controle
-                sqlRet = deleteByCtrlType(tk_ticket_ctrl.getCtrl_type(), sbWhere,db,daoObjReturn);
+                sqlRet = deleteByCtrlType(tk_ticket_ctrl, sbWhere,db,daoObjReturn);
                 //Se delete do processo "filho" OK, segue para o delete do ctrl
                 if(sqlRet != 0){
                     sqlRet = 0;
@@ -538,46 +538,59 @@ public class TK_Ticket_CtrlDao extends BaseDao implements DaoWithReturn<TK_Ticke
 
     /**
      * Deleta o processo filho do controle baseado no tipo
-     * @param ctrl_type - Tipo do control
+     * @param tk_ticket_ctrl - Control
      * @param sbWhere - Where para delete
      * @param db - Instancia do db
      * @param daoObjReturn - Obj de retorno do Dao
      * @return - Qtd de registros removidos
      * @throws Exception
      */
-    private long deleteByCtrlType(String ctrl_type, StringBuilder sbWhere, SQLiteDatabase db, DaoObjReturn daoObjReturn) throws Exception {
+    //TODO REVER POIS SE MUDAR DE VARIS ITENS PARA NENHUM, NÃO FUNCIONARÁ DEVIDAMENTE  - DESISTIR DA IDEIA DO RETORNO = 0 DAR EXCEPTION
+    //TODO POIS AGORA OS ITENS PODEM SER NULLS....
+    private long deleteByCtrlType(TK_Ticket_Ctrl tk_ticket_ctrl, StringBuilder sbWhere, SQLiteDatabase db, DaoObjReturn daoObjReturn) throws Exception {
         String ctrlTypeTable = "";
         //
-        switch (ctrl_type){
+        switch (tk_ticket_ctrl.getCtrl_type()){
             case ConstantBaseApp.TK_TICKET_CRTL_TYPE_MEASURE:
                 ctrlTypeTable = TK_Ticket_MeasureDao.TABLE;
-                //Tenta o delete
-                //Como a pk é a mesma e a relação é 1 para 1 será feito um db.delete diretamente daqui
-                daoObjReturn.setTable(ctrlTypeTable);
-                return db.delete(ctrlTypeTable,sbWhere.toString(),null);
+                //Tenta o delete se houver
+                if(tk_ticket_ctrl.getMeasure() != null) {
+                    //Como a pk é a mesma e a relação é 1 para 1 será feito um db.delete diretamente daqui
+                    daoObjReturn.setTable(ctrlTypeTable);
+                    return db.delete(ctrlTypeTable, sbWhere.toString(), null);
+                }
+                return 1;
             case ConstantBaseApp.TK_TICKET_CRTL_TYPE_ACTION:
                 //Tenta o delete
                 //Como a pk é a mesma e a relação é 1 para 1 será feito um db.delete diretamente daqui
                 ctrlTypeTable = TK_Ticket_ActionDao.TABLE;
-                daoObjReturn.setTable(ctrlTypeTable);
-                return db.delete(ctrlTypeTable,sbWhere.toString(),null);
+                if(tk_ticket_ctrl.getAction() != null) {
+                    daoObjReturn.setTable(ctrlTypeTable);
+                    return db.delete(ctrlTypeTable, sbWhere.toString(), null);
+                }
+                return 1;
             case ConstantBaseApp.TK_TICKET_CRTL_TYPE_APPROVAL:
+                int approvalDelete = 1;
                 //LUCHE - 24/07/2020
                 //Aqui ja mudou tudo, não é necessariamente 1 x 1, pode ter só aproval ou só rejeição
                 //e rejeição é uma lista...TOP
                 //Como não é 1 pra 1
                 ctrlTypeTable = TK_Ticket_ApprovalDao.TABLE;
-                daoObjReturn.setTable(ctrlTypeTable);
-                int approvalDelete = db.delete(ctrlTypeTable, sbWhere.toString(), null);
+                if(tk_ticket_ctrl.getApproval() != null) {
+                    daoObjReturn.setTable(ctrlTypeTable);
+                    approvalDelete = db.delete(ctrlTypeTable, sbWhere.toString(), null);
+                }
                 //
                 ctrlTypeTable = TK_Ticket_Approval_RejectionDao.TABLE;
-                daoObjReturn.setTable(ctrlTypeTable);
-                approvalDelete += db.delete(ctrlTypeTable, sbWhere.toString(), null);
+                if(tk_ticket_ctrl.getRejection() != null && tk_ticket_ctrl.getRejection().size() > 0 ) {
+                    daoObjReturn.setTable(ctrlTypeTable);
+                    approvalDelete += db.delete(ctrlTypeTable, sbWhere.toString(), null);
+                }
                 return approvalDelete;
             default:
-                break;
+                return 1;
         }
-        return 0;
+       // return 0;
     }
 
     /**
